feat(attachment): implement URI-based attachment system
**jp_attachment**

This crate provides a trait for handling attachments and a registry for
handling attachments in a generic way. The trait supports adding,
removing and listing attachments. Each attachment handler "owns" a
single URI scheme, and can handle attachments of that scheme.

**jp_attachment_file_content**

This crate provides an attachment handler for the full content of a
file. It supports glob patterns for file inclusion and exclusion.

More handlers will be added in the future, such as "file headers", "file
summaries", "web pages", etc.

Signed-off-by: Jean Mertz <git@jeanmertz.com>