feat(jp_id): implement custom ID system with parsing and formatting
The ID system allows for a standardized way to identify and reference
different entities in JP.

The generic format of an ID is:

```text
jp-{variant}{target_id}-{global_id}
```

Where:

- `jp` is the prefix for all IDs
- `variant` is a single character that groups IDs of the same type
  together (e.g. `c` for conversation, `m` for model, `p` for persona).
- `target_id` is a unique identifier for a given "type" (e.g. a
  conversation).
- `global_id` is a unique identifier for a specific workspace.

This allows us to print unique IDs to the terminal, which allows users
to use them in commands and scripts, or to program their editor to allow
clicking on IDs to perform specific actions.

For example, clicking on `jp-c17457886043-otvo8` in the terminal can
trigger the `jp conversation show 17457886043` command.

Note that in most cases, when using the CLI, the user only needs the
"target id" portion of the ID, which is enough to uniquely identify a
given entity, such as in the example above.

Signed-off-by: Jean Mertz <git@jeanmertz.com>