refactor(storage): more fine-grained deletion/modification tracking
Before, the storage system used a temporary directory to copy the
workspace state to upon initialization of the CLI. This was done to
ensure atomic replacement of the state directory, but this came with
downsides:

- Any storage files *not* loaded into memory (e.g. due to
  deserialization errors) were removed once the new state was persisted
  and the entire storage directory was replaced.
- When tracking the storage directory in a VCS, every run of the CLI
  would change the storage directory metadata, potentially causing
  non-relevant changes to be committed.
- The wholesale copying/replacing of the storage directory was
  inefficient.

With this commit, the storage system is no longer *atomic* (with some
additional work, it could be made to be), but we do gain other benefits:

- The workspace state is loaded from the storage directory directly,
  without the need for a temporary copy.
- Internally, modifications and deletions are tracked, to allow
  fine-grained control over what is persisted or removed.
- VCS metadata is no longer updated for every run of the CLI, only when
  a file has actually changed.

A new `TombMap` type is introduced to track deletions and modifications
to the workspace state. This is a direct copy of Rust's `HashMap`, with
minor modifications to track deletions/modifications.

Additionally, the different `*Id` types have been updated to correctly
convert between paths/filenames and IDs.

Signed-off-by: Jean Mertz <git@jeanmertz.com>