feat(conversation): add support for private conversations (#1)
Private conversations are a crucial feature to make JP useable for any
situation. Most of the time, both your work and the conversations you
had with JP to get that work finished are valuable to the team, so you
want to commit both artifacts to the VCS. However, sometimes you are
stuck and just need a bit of help solving a problem, but the
conversation itself provides no value to others in the team.

In these situations, you may want to keep the conversation private, so
that the conversation is stored locally, outside of the workspace,
accessible only to you. This way, you can still use JP to solve the
problem, but the conversation is not visible to others in the team.

User-facing changes:

- `jp conversation --new --private` creates a new private conversation.
- Private converations are stored in the local storage only, not in the
  workspace storage.
- `jp conversation ls --private` shows only private conversations.
- `jp conversation show` shows the private status of a conversation.

Signed-off-by: Jean Mertz <git@jeanmertz.com>