feat(conversation): edit conversation details
Specific conversation details can now be modified using the
`conversation edit` subcommand.

```sh
$ jp conversation edit
Edit conversation details

Usage: jp conversation edit [OPTIONS] <ID|--private [<PRIVATE>]|--title [<TITLE>]>

Arguments:
  [ID]  Conversation ID to edit. Defaults to the active conversation if not specified

Options:
      --private [<PRIVATE>]  Toggle the private flag of the conversation [possible values: true, false]
      --title [<TITLE>]      Edit the title of the conversation
```

For `--private`, the flag can be `true` or `false`, but can also be used
as a toggle using `--private` without a value, in which case a private
conversation becomes public and vice versa.

For `--title`, a title can be provided as an argument, or if none is
given, *in a future version*, the LLM will be used to generate a title.

Technical details:

- Add new `edit` subcommand to modify conversation properties
- Support toggling private flag and updating title
- Conversation directory cleanup when moving between private/public
- Temporary workaround to change detection in `TombMap::iter_mut`

Signed-off-by: Jean Mertz <git@jeanmertz.com>