fix(cli): improve explicit model override handling
Previously, an explicit model override would change the provider and
model name of whatever model was currently loaded. This could cause
issues if the loaded model was (for example) a reasoning model, but the
explicit model was not. In this case, the LLM API could return an error
indicating that reasoning was not supported for the given mode.

With this change, if we can find a pre-configured model with the same
provider and model name of the explicit model, we will use that,
otherwise we will fall back to the default (non-reasoning) model
configuration with the provided provider and model name.

This can technically still result in an error if a model needs specific
configurations that are not the default, but this should be rare, and in
such cases the user can create a model configuration file with the
proper model attributes.

Signed-off-by: Jean Mertz <git@jeanmertz.com>