fix(editor): clean up `QUERY_MESSAGES.md` file when empty (#11)
When the user saves an empty query file, we already ignored the query
and exited the command early. Now we also remove the empty file to clean
up after ourselves.

(note: the file isn't "empty", it contains the context of the earlier
messages in the conversation, but if there is no non-whitespace text
before the "marker line", then we consider the file empty)

Signed-off-by: Jean Mertz <git@jeanmertz.com>