feat(conversation): automated title generation
With this commit, newly created conversations will automatically
generate a title based on the initial conversation query.

To enable this feature, a new *background task scheduling* system has
been added in the new `jp_task` crate.

When a task is spawned, it is run to completion in the background while
the CLI continues to run. When the CLI is about to exit, the task gets
to sync its state with the workspace, ensuring that the task's results
are persisted.

Any still-running tasks are asked to stop gracefully using a
cancellation token, but they only have a limited amount of time to stop,
before they are forcibly terminated, to avoid blocking the CLI.

Currently two background tasks are implemented:

- TitleGeneratorTask: Generates a conversation title based on the
  conversation's content.
- StatelessTask: Runs a generic future, without any state to sync with
  the workspace.

Signed-off-by: Jean Mertz <git@jeanmertz.com>