perf: improveo overall performance of CLI commands
This commit focuses on improving the performance of running JP. The
biggest gain is in attachment handling, which went from ~10s for the JP
repo to < 1s. This was mostly achieved (a) parallel traversing, and (b)
by not traversing non-included directories, instead of first collecting
*all* files in a workspace, and *then* filtering them (this resulted in
collecting all files in e.g. `.git` or `target` directories).

Additionally, the project has seen its color shifting from blue to
red[1] as more functions are now async, allowing for faster resolution
of tasks, especially during HTTP calls to LLM providers.

[1]: https://journal.stuffwithstuff.com/2015/02/01/what-color-is-your-function/

Signed-off-by: Jean Mertz <git@jeanmertz.com>