refactor(config): allow disabling inheritance from CLI (#32)
Before this change, `--cfg inherit=false` would disable *file-based*
inheritance from the CLI, but this (a) did not work, and (b) was limited
in in scope.

It did not work because CLI flags override `Config` values *after* files
and environment variables are loaded, so even if it would set `inherit`
to `false`, the files were already loaded, including using the default
inheritance behavior.

It was limited in scope, because even if the above were fixed, it would
not allow disabling inheritance from the root config file, nor from
environment variables.

Now, `--cfg inherit=false` will disable inheritance from the CLI
"downward", meaning, it will *only* take into account configuration
options set via the CLI, and not from files or environment variables.

(note, this commit also changes the `--config` flag to `--cfg`. This was
already `--cfg` before, but got inadvertently changed to `--config` in a
previous commit.)

Signed-off-by: Jean Mertz <git@jeanmertz.com>