feat(query): support structured output (#36)
The new `--schema` flag allows you to specify a JSON schema for the
output of a `query` command. If a schema is provided, the assistant will
conform its output to the schema, and the output will printed as JSON.

The JSON schema can be provided as a string, or a `@`-prefixed path to a
JSON file.

If the output of the command is piped or redirected, the JSON output
will be printed in compact form, otherwise it will be pretty-printed.
This allows for post-processing of the output using tools like `jq`.

For now, structured output is *not streamed*, meaning its output is
printed after the assistant has finished generating the response. This
will change in the future, by adding a generic `--[no-]stream` flag.

Signed-off-by: Jean Mertz <git@jeanmertz.com>