refactor(config)!: empty strings in config setters unset optionals (#41)
Before this commit, an optional configuration value that was set, could
not be unset with the `--cfg` flag. This commit fixes that.

For example:

```sh
jp query --cfg conversation.context= "Is this thing on?"
```

The above command will now set the `Conversation::context` configuration
field to `None`.

This is useful if e.g. an optional field is set in `.jp/config.toml` but
you want to clear it for a specific conversation.

Signed-off-by: Jean Mertz <git@jeanmertz.com>