test: Improve VCR-based snapshot testing with insta
The `jp_test` crate's VCR harness has been refactored to use insta for
snapshot testing.

The return value of the `cassette` method is now used to generate a
snapshot of the test result. This makes it easy to record a cassette
without having to manually assert the result.

The snapshots are auto-accepted, because we otherwise run into the
problem where a new snapshot causes a test failure (which is how insta
snapshot testing works by default), which then causes the `httpmock`
crate to abort recording the request, which then means even if the
snapshot is manually accepted, when re-running the test, no http
recording exists, and so another request is made, which invalidates the
earlier accepted snapshot (because of the non-deterministic output of
the LLM APIs to the same request). Snapshots still need to be reviewed
when committing the snapshots to the repo.

The approach eliminates boilerplate assertion code and provides a good
basis for expanding our test suite without overcomplicated test setups.

Signed-off-by: Jean Mertz <git@jeanmertz.com>