refactor(workspace): Rename `LocalState` to `UserState` (#64)
Previously, the `State` struct had fields named `workspace` and `local`,
where `local` actually referred to user metadata. This led to confusing
naming and misuse. With this change, `State.local` now holds the
workspace state (it's state local to the workspace), and the former
`LocalState` type is renamed to `UserState`, stored in `State.user`. All
references to the old fields and type have been updated accordingly.

Signed-off-by: Jean Mertz <git@jeanmertz.com>