refactor: Replace `backoff` with `backon` for retry logic (#65)
This commit migrates the retry implementation from `backoff` to
`backon`, as the former is deprecated and no longer maintained.

The `async-anthropic` dependency is also updated to a specific
(unreleased) version which includes a few fixes and also swaps `backoff`
for `backon` in its retry logic. See the following PR for more details:
https://github.com/bosun-ai/async-anthropic/pull/14

This work is in preparation of adding `cargo-deny` to the project.

Signed-off-by: Jean Mertz <git@jeanmertz.com>