build(justfile): Add commit recipe to justfile
This adds a commit recipe to the `justfile` that automates commit
message creation. Invoking `just commit` will run `jp query` in a mode
that generates a one-off commit message response from the assistant
based on the current staged changes. For now, `sed` is used to strip
ANSI escape codes, but in the future there will be an option in `jp` to
not include the ANSI escape codes in the response. Finally, the output
is piped to `git commit`, opening the user's editor with the commit
message pre-populated.

This streamlines the commit process and ensures consistent formatting.

Signed-off-by: Jean Mertz <git@jeanmertz.com>