build: Add `cargo-deny` to the build process (#77)
In this commit, we introduce `deny.toml` to configure the deny crate for
dependency auditing. The configuration ignores advisory
RUSTSEC-2024-0436 for paste, specifies a list of approved licenses, and
denies unknown registries and git sources.

To pass the `cargo deny` check, we bump the tokio dependency from 1.44.1
to 1.45.0. We add an Apache License 2.0 `LICENSE` file to clearly state
the project license and improve compliance.

Signed-off-by: Jean Mertz <git@jeanmertz.com>