fix(llm): Correctly set context window size for Ollama
Previously, the code incorrectly used `max_tokens` to set Ollama's
`num_ctx` option. This was problematic because `max_tokens` typically
refers to the maximum number of tokens to generate in a response, while
`num_ctx` controls the model's context window size - the total amount of
context the model can process.

This change introduces a dedicated `context_window` parameter that can
be specified in the model parameters. When set, this value will be used
to configure Ollama's context window size, allowing users to force
Ollama to use a larger context window than what it would automatically
determine based on the machine's available resources.

Signed-off-by: Jean Mertz <git@jeanmertz.com>