feat(llm): Improved reasoning support (#105)
For the best possible conversational experience, it is important to
include reasoning data in historical turns. However, this can
significantly increase token usage. Some providers allow you to embed
the reasoning data, but they strip this data out on their end, reducing
the token count. For this to work, we need to strictly adhere to the API
specification of the providers.

For example, OpenAI requires one of two things:

- Either you enable stateful conversations, which means data is stored
and retained on the servers of OpenAI,

- Or you manually build up the conversation context, which means no data
needs to be stored on the servers.

We use the latter approach, to avoid additional data storage, but this
means we need to send the data in the exact shape as the API expects it.

Since every provider has different API expectations, we now embed an
opaque `metadata` map in the `AssistantMessage` structure. This map
stores keys with arbitrary json values. The map is populated by the
providers when response is received, and read again when a new request
is built.

In the case of OpenAI, we store a `reasoning` data blob in the metadata,
which we then embed in the request to include reasoning tokens in the
conversation.

This also means we removed the old free-form `Thinking` implementation.
This wasn't used anywhere as this was from an earlier version of JP in
which we supported sending reasoning requests to one provider and the
regular conversation to another, embedding the reasoning data of the
first provider in the second provider's request. Now that more/all
providers support reasoning, the importance of this feature is reduced.
We can re-introduce it in the future, if we need it.

This commit does not yet implement similar strategies for other
providers. We will do that in follow up commits.

Signed-off-by: Jean Mertz <git@jeanmertz.com>