fix(llm, anthropic): Disable reasoning when tool choice forces tool use (#149)
Anthropic's API does not support reasoning when tool choice is forced,
resulting in an error when this was the case. The Anthropic provider now
properly detects when `tool_choice` forces tool usage and automatically
disables reasoning in these cases. An `INFO` log message is emitted to
inform the user of this behavior.

Signed-off-by: Jean Mertz <git@jeanmertz.com>