feat(cli): Show persona and model info in query editor (#152)
When editing queries in the CLI, the editor now displays the current
persona (if set) and model information right below the start of the
metadata part of the file. This provides users with immediate context
about their conversation settings while composing their queries, making
it easier to understand which model and persona configuration will be
used for the interaction.

Signed-off-by: Jean Mertz <git@jeanmertz.com>