build: Add `install` target and update coverage generation
Added a new `install` target that depends on `_install-jp` to provide
a convenient way to install the CLI tool. The `commit` target is now
organized under a `git` group for better categorization of justfile
recipes.

Updated the `coverage-ci` target to disable `cfg-coverage` settings in
`llvm-cov`, which (hopefully) improves build times on CI.

Signed-off-by: Jean Mertz <git@jeanmertz.com>