feat(mcp): Add MCP security and display configuration options (#173)
Users can now configure how MCP tool calls are handled during a
conversation. Both MCP servers and tools can be configured. Default
values can be set for all servers and tools, with the ability to
override them on a per-server or per-tool basis.

The following server settings are available:

- `enable` — This setting propagates to all tools of the server,
allowing to disable all server tools with a single switch. By default,
all servers are enabled.

- `binary_checksum` — The checksum of a "stdio" server binary will be
verified before a tool call is made, preventing malicious actors from
injecting undesired code into the server. The
`binary_checksum.algorithm` setting can be set to `sha1` or `sha256`,
while the `checksum.value` setting is set to the expected checksum.

- `tools` — A map of tool names to tool configurations.

Each tool configuration has the following settings:

- `enable` — If a tool is disabled, it will be hidden from the assistant
and will not be called during a conversation.

- `run` — How to handle a tool call run. Either `always` to allow
running the tool without confirmation, `ask` (default) to ask for
confirmation before running the tool, or `edit` to open an editor to
edit the tool call parameters before running it.

- `result` — What to do before delivering the tool call result. Same
options as for the run mode.

- `style` — How to style the tool call results. The following styles are
available:

- `inline_results` — How to print the tool call results in the terminal.
Either `off` to hide the results in the terminal, `full` (default) to
print the full results inline, or an integer to print the first `N`
lines of the results inline. Regardless of this setting, the full
results will be delivered to the assistant, and a link to a file
containing the full results is printed to the terminal.

- `results_file_link` — How to link to the file containing the tool call
results. Either `off` to hide the link, `full` to print a `file://` link
in the terminal, or `osc8` to link to the file using the
[OSC8](https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda)
format.

---------

Signed-off-by: Jean Mertz <git@jeanmertz.com>