fix(mcp): Expose tool parameter type information (#214)
Previously, the type property of a tool's parameters was inadvertently
filtered out and not exposed to the language model. This limited the
context the model had about the expected data type for each argument.

It also resulted in the Google provider from returning errors, as it is
more strict about the schema definition of tools. This strictness also
made it impossible to add the `exmamples` property to the schema, those
details are now moved to the `description` property.

Signed-off-by: Jean Mertz <git@jeanmertz.com>