fix(config): Inherit wildcard configs for MCP servers and tools (#224)
Previously, during configuration parsing, after we had finished
wildcard-inheritance for defined servers and tools, we would throw away
the wildcard configurations, as they were no longer needed (or so I
thought).

Turns out, we DO need them. Specifically, in a scenario where you define
e.g.:

```toml
[servers.github.tools.*]
enable = false
```

This would (or rather, _should_) disable all tools for the `github`
server. However, because we _merged_ global defaults with _specific
servers and tools_, and since there **is not specific server or tool
defined in this example**, no merging would happen, and the wildcard
configuration got thrown out.

What I failed to realize was that the merging-at-parse-time works in a
case such as:

```toml
[servers.github.tools.*]
enable = false

[servers.github.tools.get_readme]
```

In the above case, the configuration contains a specific tool
configuration (even if empty), and inherits the wildcard configuration.
But in the first case, no specific tool configuration exists.

Still though, at runtime, if we were to call the `get_readme` tool, we
would still want the default from the first example to be applied to a
non-existing tool (or server) configuration. That's what this commit
rectifies.

Because we merge the existing configurations at startup, we don't need
to do any merging at runtime. If a configuration exists for a specific
server or tool, it has already been merged with the wildcard ones, so we
can just return it as-is, if none exists, we can just return the
wildcard configuration, or the `default()` impl.

---------

Signed-off-by: Jean Mertz <git@jeanmertz.com>